organized webpack scripts into new 'webpack' folder
